requirements.txt
azure-core==1.4.0
azure-storage-blob==12.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.9
isodate==0.6.0
msrest==0.6.13
numpy==1.18.3
oauthlib==3.1.0
pandas==1.0.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.9
xlrd==1.2.0